# public/openapi.yaml
openapi: 3.0.3
info:
  title: REVA Mail API
  description: |
    API REST Lengkap untuk REVA Mail - Layanan Email
    
    ## Fitur
    - Buat alamat email
    - Terima dan kelola email
    - Pemeriksaan email real-time
    - Simulasi email untuk pengujian
    - Pemrosesan email massal
    
    ## Otentikasi
    API ini menggunakan proteksi CSRF. Sertakan header `X-CSRF-TOKEN` dalam permintaan Anda.
    
    ## Batas Rate
    - 60 permintaan per menit per IP
    - 10 pembuatan email per jam
    - 20 simulasi per jam
    
  version: 1.0.0
  contact:
    name: REVA Mail API Support
    url: https://revacantik.my.id
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://revacantik.my.id/api/v1
    description: Server produksi
  - url: http://127.0.0.1:8000/api/v1
    description: Server pengembangan
paths:
...
  /temp-emails/{emailId}:
    delete:
      summary: Delete temporary email
      description: Permanently deletes a temporary email and all its received messages
      tags:
        - Email Management
      security:
        - csrfAuth: []
      parameters:
        - name: emailId
          in: path
          required: true
          description: The ID of the temporary email
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Email deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: "Temporary email deleted successfully"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '419':
          $ref: '#/components/responses/CSRFError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /temp-emails/{emailId}/email/{messageId}:
    get:
      summary: Get specific email
      description: Retrieves details of a specific email message and marks it as read
      tags:
        - Inbox Operations
      parameters:
        - name: emailId
          in: path
          required: true
          description: The ID of the temporary email
          schema:
            type: integer
            example: 1
        - name: messageId
          in: path
          required: true
          description: The ID of the email message
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Email retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    csrfAuth:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: CSRF token for Laravel applications
  schemas:
...
    ReceivedEmail:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the received email
          example: 1
        from_email:
          type: string
          format: email
          description: Sender's email address
          example: "noreply@example.com"
        sender_name:
          type: string
          description: Sender's display name
          example: "Example Service"
        subject:
          type: string
          description: Email subject
          example: "Welcome to our service!"
        body:
          type: string
          description: Email body content
          example: "Thank you for signing up..."
        verification_code:
          type: string
          nullable: true
          description: Verification code if present
          example: "123456"
        link:
          type: string
          format: uri
          nullable: true
          description: Verification or action link if present
          example: "https://example.com/verify?token=abc123"
        received_at:
          type: string
          format: date-time
          description: Email received timestamp (ISO 8601)
          example: "2026-03-14T08:35:20.000000Z"
        is_read:
          type: boolean
          description: Whether the email has been read
          example: false
...
    EmailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/ReceivedEmail'
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"
        errors:
          type: object
          nullable: true
          additionalProperties:
            type: array
            items:
              type: string
          example:
            field_name: ["Validation error message"]
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Temporary email not found or expired"
    CSRFError:
      description: CSRF token mismatch
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "CSRF token mismatch"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Internal server error"
...